prototype for function that checks if a move leaves the king in check
The function will return true if the move leaves the king in check. The switchstatement will check once if the move leaves the king in check at the very end. The function is not in each switch statement, but if a move is deemed valid (no obstruction in path, doesn't take friendly piece etc), the switch statement will break and then check for whether or not the move leave the king in check.